requirements.txt
fastapi
uvicorn
streamlit
openai
langchain
transformers
tiktoken
numpy
pandas
sqlalchemy
google-api-python-client
python-dotenv
sentence-transformers
faiss-cpu
requests
ollama
pdfplumber
python-docx
python-pptx
pytesseract
pillow
pdf2image
opencv-python
python-multipart
sympy
pix2text